+++ src/notes.test.ts
@@ -1,5 +1,9 @@
 import {
+	getAttributeByPath,
+	getAttributesByPath,
+	getAttributeValueByPath,
 	getCoverUrl,
+	getLabelValueByPath,
 	getSortableAttributeValue,
 	getSortableGroupName,
 	getSortableTitle,
@@ -7,6 +11,132 @@ import {
 	sortNotes,
 } from "collection-views/notes";
 import { MockApi, MockNoteShort } from "collection-views/test";
+
+const attributeNote = new MockNoteShort({
+	noteId: "1",
+	title: "Title",
+	attributes: [
+		{ type: "label", name: "test", value: "Label" },
+		{ type: "relation", name: "test", value: "Relation" },
+		{ type: "label", name: "label", value: "2" },
+		{ type: "relation", name: "relation", value: "2" },
+		{ type: "relation", name: "relation", value: "3" },
+		{ type: "relation", name: "relation", value: "bad" },
+	],
+});
+const relatedNotes = [
+	new MockNoteShort({
+		noteId: "2",
+		title: "Related note 1",
+		attributes: [
+			{ type: "label", name: "label", value: "Label 1" },
+			{ type: "label", name: "label", value: "Label 2" },
+		],
+	}),
+	new MockNoteShort({
+		noteId: "3",
+		title: "Related note 2",
+		attributes: [{ type: "label", name: "label", value: "Label 3" }],
+	}),
+];
+
+describe("getAttributesByPath", () => {
+	beforeEach(() => {
+		new MockApi({ notes: relatedNotes });
+	});
+
+	test.each([
+		["returns ID for $id", "$id", "1"],
+		["returns ID for $noteId", "$noteId", "1"],
+		["returns type", "$type", "text"],
+		["returns content type", "$mime", "text/html"],
+		["returns title", "$title", "Title"],
+		["returns content length", "$contentSize", "1000"],
+		["returns date created", "$dateCreated", "2020-01-02 03:04:05.678Z"],
+		["returns date modified", "$dateModified", "2020-02-03 04:05:06.789Z"],
+	])("%s", async (_, path, expected) => {
+		const attributes = await getAttributesByPath(attributeNote, path);
+		expect(attributes).toHaveLength(1);
+		expect(attributes[0].type).toBe("label");
+		expect(attributes[0].value).toBe(expected);
+	});
+
+	test.each([
+		["returns empty array for empty name", "", []],
+		["returns empty array for invalid property", "$bad", []],
+		["returns attributes", "test", ["Label", "Relation"]],
+		[
+			"returns related notes' attributes",
+			"relation.label",
+			["Label 1", "Label 2", "Label 3"],
+		],
+		["returns empty array for missing attribute", "bad", []],
+		["returns empty array for attributes of labels", "label.label", []],
+		[
+			"returns empty array for attributes of missing attributes",
+			"bad.bad",
+			[],
+		],
+		[
+			"returns related notes' properties",
+			"relation.$title",
+			["Related note 1", "Related note 2"],
+		],
+		[
+			"returns empty array for missing attributes for related notes",
+			"relation.bad",
+			[],
+		],
+	])("%s", async (_, path, expected) => {
+		const attributes = await getAttributesByPath(attributeNote, path);
+		expect(attributes.map((attribute) => attribute.value)).toEqual(
+			expected
+		);
+	});
+});
+
+describe("getAttributeByPath", () => {
+	test("returns first attribute", async () => {
+		const attribute = await getAttributeByPath(attributeNote, "test");
+		expect(attribute).not.toBeNull();
+		expect(attribute?.value).toBe("Label");
+	});
+
+	test("returns null if not found", async () => {
+		const attribute = await getAttributeByPath(attributeNote, "bad");
+		expect(attribute).toBeNull();
+	});
+});
+
+describe("getAttributeValueByPath", () => {
+	test.each([
+		["returns value of first attribute", "test", "Label"],
+		["returns empty string if not found", "bad", ""],
+	])("returns %s", async (_, path, expected) => {
+		const value = await getAttributeValueByPath(attributeNote, path);
+		expect(value).toBe(expected);
+	});
+});
+
+describe("getLabelValueByPath", () => {
+	test("returns value of first label", async () => {
+		const note = new MockNoteShort({
+			attributes: [
+				{ type: "relation", name: "test", value: "Relation" },
+				{ type: "label", name: "test", value: "Label 1" },
+				{ type: "label", name: "test", value: "Label 2" },
+			],
+		});
+
+		const value = await getLabelValueByPath(note, "test");
+		expect(value).toBe("Label 1");
+	});
+
+	test("returns empty string if not found", async () => {
+		const value = await getLabelValueByPath(attributeNote, "relation");
+		expect(value).toBe("");
+	});
+});
 
 describe("getCoverUrl", () => {
 	test.each([
@@ -22,11 +152,11 @@ describe("getCoverUrl", () => {
 });
 
 describe("groupNotes", () => {
-	test("returns an empty array if no notes", async () => {
+	test("returns empty array if no notes", async () => {
 		expect(await groupNotes([], "group")).toEqual([]);
 	});
 
-	test("returns groups otherwise", async () => {
+	test("returns groups for attribute", async () => {
 		const relatedNotes = [
 			new MockNoteShort({ noteId: "1", title: "Note 1" }),
 			new MockNoteShort({
@@ -107,6 +237,43 @@ describe("groupNotes", () => {
 			},
 			{ name: "z", relatedNote: null, notes: [notes[2]] },
 			{ name: undefined, relatedNote: null, notes: [notes[0], notes[1]] },
+		]);
+	});
+
+	test("returns groups for related note's attribute", async () => {
+		const notes = [
+			new MockNoteShort({
+				attributes: [
+					{ type: "relation", name: "relation", value: "1" },
+				],
+			}),
+			new MockNoteShort({
+				attributes: [
+					{ type: "relation", name: "relation", value: "2" },
+				],
+			}),
+		];
+		const relatedNotes = [
+			new MockNoteShort({
+				noteId: "1",
+				attributes: [
+					{ type: "label", name: "label", value: "Label 1" },
+				],
+			}),
+			new MockNoteShort({
+				noteId: "2",
+				attributes: [
+					{ type: "label", name: "label", value: "Label 2" },
+				],
+			}),
+		];
+
+		new MockApi({ notes: relatedNotes });
+
+		const groups = await groupNotes(notes, "relation.label");
+		expect(groups).toEqual([
+			{ name: "Label 1", relatedNote: null, notes: [notes[0]] },
+			{ name: "Label 2", relatedNote: null, notes: [notes[1]] },
 		]);
 	});
 });
@@ -239,6 +406,10 @@ describe("sortNotes", () => {
 		}),
 	];
 
+	beforeEach(() => {
+		new MockApi({ notes: relatedNotes });
+	});
+
 	test.each([
 		[
 			"sorts notes by title",
@@ -259,38 +430,37 @@ describe("sortNotes", () => {
 		[
 			"sorts notes by text ascending, then title",
 			textNotes,
-			[{ name: "text", descending: false }],
+			[{ path: "text", descending: false }],
 			["5", "2", "1", "3", "4"],
 		],
 		[
 			"sorts notes by text descending, then title",
 			textNotes,
-			[{ name: "text", descending: true }],
+			[{ path: "text", descending: true }],
 			["4", "2", "1", "3", "5"],
 		],
 		[
 			"sorts notes by number ascending, then title",
 			numericNotes,
-			[{ name: "number", descending: false }],
+			[{ path: "number", descending: false }],
 			["7", "2", "1", "3", "5", "4", "6"],
 		],
 		[
 			"sorts notes by number descending, then title",
 			numericNotes,
-			[{ name: "number", descending: true }],
+			[{ path: "number", descending: true }],
 			["6", "4", "5", "2", "1", "3", "7"],
 		],
 		[
 			"sorts notes by multiple attributes",
 			multipleNotes,
 			[
-				{ name: "one", descending: false },
-				{ name: "two", descending: true },
+				{ path: "one", descending: false },
+				{ path: "two", descending: true },
 			],
 			["3", "5", "2", "4", "1"],
 		],
 	])("%s", async (_, notes, sortAttributes, expected) => {
-		new MockApi({ notes: relatedNotes });
 		await sortNotes(notes, sortAttributes);
 		expect(notes.map((note) => note.noteId)).toEqual(expected);
 	});
@@ -333,54 +503,56 @@ describe("getSortableGroupName", () => {
 });
 
 describe("getSortableAttributeValue", () => {
+	const notes = [
+		new MockNoteShort({
+			attributes: [
+				{ type: "label", name: "label", value: "  Value 1  " },
+				{ type: "label", name: "label", value: "Value 2" },
+				{ type: "relation", name: "relation", value: "1" },
+				{ type: "relation", name: "relation", value: "2" },
+				{ type: "relation", name: "sortable", value: "2" },
+				{ type: "relation", name: "notFound", value: "  Bad  " },
+			],
+		}),
+		new MockNoteShort({
+			noteId: "1",
+			title: "  Title 1  ",
+			attributes: [
+				{ type: "label", name: "label", value: "  Related Label 1  " },
+			],
+		}),
+		new MockNoteShort({
+			noteId: "2",
+			title: "  Title 2  ",
+			attributes: [
+				{ type: "label", name: "sortableTitle", value: "  Sortable  " },
+				{ type: "label", name: "label", value: "Related Label 2" },
+			],
+		}),
+	];
+
+	beforeEach(() => {
+		new MockApi({ notes });
+	});
+
 	test.each([
+		["returns label's value", "label", "value 1"],
+		["returns related note's title", "relation", "title 1"],
+		["returns related note's sortableTitle", "sortable", "sortable"],
 		[
-			"returns a label's value",
-			[],
-			[
-				{ type: "label", name: "test", value: "  Value 1  " },
-				{ type: "label", name: "test", value: "  Value 2  " },
-			],
-			"value 1",
-		],
-		[
-			"returns a related note's title",
-			[new MockNoteShort({ noteId: "1", title: "  Title  " })],
-			[
-				{ type: "relation", name: "test", value: "1" },
-				{ type: "relation", name: "test", value: "2" },
-			],
-			"title",
-		],
-		[
-			"returns a related note's sortableTitle",
-			[
-				new MockNoteShort({
-					noteId: "1",
-					title: "  Title  ",
-					attributes: [
-						{
-							type: "label",
-							name: "sortableTitle",
-							value: "  Sortable Title  ",
-						},
-					],
-				}),
-			],
-			[{ type: "relation", name: "test", value: "1" }],
-			"sortable title",
-		],
-		[
-			"returns a relation's value if related note not found",
-			[],
-			[{ type: "relation", name: "test", value: "  Bad  " }],
+			"returns relation's value if related note not found",
+			"notFound",
 			"bad",
 		],
-		["returns an empty string if attribute not found", [], [], ""],
-	])("%s", async (_, notes, attributes, expected) => {
-		new MockApi({ notes });
-		const note = new MockNoteShort({ attributes });
-		expect(await getSortableAttributeValue(note, "test")).toBe(expected);
+		[
+			"returns related note's label's value",
+			"relation.label",
+			"related label 1",
+		],
+		["returns empty string if attribute not found", "bad", ""],
+	])("%s", async (_, path, expected) => {
+		const value = await getSortableAttributeValue(notes[0], path);
+		expect(value).toBe(expected);
 	});
 });