+++ CHANGELOG.md
@@ -2,6 +2,29 @@
 
 ## 1.2.0 - Unreleased
 
+- Add support for using properties of a note.
+  - Wherever an attribute name can be specified, there are now some special names (prefixed with `$`) which refer to a note's properties instead of its user-defined attributes.
+  - `$id` and `$noteId` are the note's ID.
+  - `$type` is the note type (for example, `text`).
+  - `$mime` is the note's content type (for example, `text/html`).
+  - `$title` is the note's title.
+  - `$contentSize` is the size of the note's content in bytes.
+  - `$dateCreated` is the note's creation date and time in UTC and RFC 3339 format (`YYYY-MM-DD hh:mm:ss.sssZ`).
+  - `$dateModified` is the note's modification date and time in UTC and RFC 3339 format.
+- Add support for using attributes of notes targeted by a note's relations, similar to that supported by [Trilium's search engine](https://github.com/zadam/trilium/wiki/Search#advanced-use-cases).
+  - Wherever an attribute name can be specified, a "path" can now be specified instead.
+  - A path consists of one or more names separated by a period (`.`). The last name in the path must be an attribute name or a property name. All other names in the path must be relation names.
+  - For example:
+    - `name` would find attributes named `name` defined on a note.
+    - `employee.name` would find attributes named `name` defined on all notes targeted by the `employee` relation defined on a note.
+    - `company.employee.name` would find attributes named `name` defined on all notes targeted by the `employee` relation defined on all notes targeted by the `company` relation defined on a note.
+- Add tokens to `#query` for substituting the Render Note's ID and attributes into the search query.
+  - `$id` and `$noteId` will be replaced with the Render Note's ID.
+  - `$renderNote.name` will be replaced with the value of the first attribute found for the Render Note (or an empty string if not found). `name` can be the name of an attribute, a property, or a related note's attribute.
+- `#attribute` now supports properties (`#attribute="$dateModified"`) and attributes of related notes (`#attribute=relation.label`).
+- `#groupBy` now supports properties (`#groupBy="$type"`) and attributes of related notes (`#groupBy=relation.label`).
+- `#sort` now supports properties (`#sort="$dateModified"`) and attributes of related notes (`#sort=relation.label`).
+- `progressBar` now supports properties (`#attribute="count,progressBar=$contentSize"`) and attributes of related notes (`#attribute="count,progressBar=relation.total"`).
 - Add a `separator` attribute setting for controlling how multiple values for a single attribute are separated:
   - `separator=newline` inserts a newline between values, resulting in one value per line.
   - `separator=comma` inserts a comma and space between values.
@@ -12,6 +35,7 @@
 - Escape sequences are now supported in attribute setting values using a backtick as the escape character: <code>``</code> and <code>\`,</code>. This allows for using a comma in settings that accept arbitrary text such as `header`.
 - Add margin around all views to better align the edges of views when using Trilium's default themes. This can be changed using the `--collection-view-margin` CSS variable.
 - Fix "ResizeObserver loop limit exceeded" errors occurring in console when the note content area is resized.
+- Fix `#query` tokens not escaping backslashes and double quotes in values.
 - Fix `boolean` checkbox styles not being applied in Trilium 0.46.
 
 ## 1.1.0 - 2021-07-21